% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbts-package.R
\docType{package}
\name{lbts-package}
\alias{lbts}
\alias{lbts-package}
\title{lbts: Miscellaneous time series functions and datasets}
\description{
Miscellaneous time series functions and datasets.
}
\author{
\strong{Maintainer}: Luca Braglia \email{lbraglia@gmail.com}

}
\keyword{internal}
