% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fmri}
\alias{fmri}
\title{fmri data}
\usage{
fmri(pt = 18, brain_region = 64, test_retest = 1)
}
\description{
subjects are at rest
}
\details{
24 subjects
70 brain region for each subject series
2 set of test: test=1, retest=2

404 misllisecond time series
}
